refactor : rule 관리 객체 생성 - 체스룰을 관리하는 객체를 생성하여 game의 책임을 나누었습니다.
